Support Wacom mode, fix external monitor bug
If you arrange your xorg.conf files so that the N-trig digitizer uses the Wacom driver, you get access to all of the buttons on the SP3 pen, but it becomes three separate devices, which breaks this script. I spent the day updating the script to loop through the multiple devices.

When you connect an external monitor, the pen and touch area stretches out to fill the new screen space, which is useless since it is not a touchscreen. So you can use MapToOutput to constrain this area to the primary display (eDP1 in xrandr). Unfortunately, it does it again when you disconnect the monitor, so i configured the script to run this command every time the number of monitors changes.

Furthermore, when an external monitor is connected, we are doing nothing, so we might as well not poll every fifth of a second. Added a time delay.